+++ src/UI/Core/TextArea.h
@@ -37,5 +37,5 @@ private:
 private:
 	StringList	mFormattedList;
 
-	int			mNumLines;
+	size_t		mNumLines;
 };
